Add chunk_data function to split data into chunks
This commit adds a new function called `chunk_data` that takes in a list of data and a chunk size, and splits the data into chunks of the specified size. The chunks are then returned as a list. This function can be used to process large datasets in smaller, more manageable chunks.